+++ docs/source/conf.py
@@ -9,6 +9,7 @@ extensions = [
     "sphinx.ext.doctest",
     "sphinx.ext.autodoc",
     "sphinx.ext.intersphinx",
+    "sphinx_rtd_theme",
 ]
 
 # use index.rst instead of contents.rst
@@ -24,7 +25,11 @@ author = "Open Microscopy Environment"
 # use in refs e.g:
 # :ref:`comparison manual <python:comparisons>`
 intersphinx_mapping = {
-    "python": ("https://docs.python.org/", None),
+    "python": ("https://docs.python.org/3", None),
     "numpy": ("https://numpy.org/doc/stable/", None),
     "zarr": ("https://zarr.readthedocs.io/en/stable/", None),
+    "dask": ("https://docs.dask.org/en/stable", None),
 }
+
+# https://github.com/readthedocs/sphinx_rtd_theme
+html_theme = "sphinx_rtd_theme"