+++ ome_zarr/cli.py
@@ -1,8 +1,9 @@
 """Entrypoint for the `ome_zarr` command-line tool."""
+
 import argparse
 import logging
 import sys
-from typing import List, Union
+from typing import Union
 
 from .csv import csv_to_zarr
 from .data import astronaut, coins, create_zarr
@@ -80,7 +81,7 @@ def csv_to_labels(args: argparse.Namespace) -> None:
     csv_to_zarr(args.csv_path, args.csv_id, args.csv_keys, args.zarr_path, args.zarr_id)
 
 
-def main(args: Union[List[str], None] = None) -> None:
+def main(args: Union[list[str], None] = None) -> None:
     """Run appropriate function with argparse arguments, handling errors."""
     parser = argparse.ArgumentParser()
     parser.add_argument(