+++ ome_zarr/reader.py
@@ -3,11 +3,11 @@
 import logging
 import math
 from abc import ABC
-from typing import Any, Dict, Iterator, List, Optional, Type, Union, cast, overload
+from collections.abc import Iterator
+from typing import Any, Optional, Union, cast, overload
 
 import dask.array as da
 import numpy as np
-from dask import delayed
 
 from .axes import Axes
 from .format import format_from_version
@@ -24,25 +24,25 @@ class Node:
     def __init__(
         self,
         zarr: ZarrLocation,
-        root: Union["Node", "Reader", List[ZarrLocation]],
+        root: Union["Node", "Reader", list[ZarrLocation]],
         visibility: bool = True,
         plate_labels: bool = False,
     ):
         self.zarr = zarr
         self.root = root
-        self.seen: List[ZarrLocation] = []
-        if isinstance(root, Node) or isinstance(root, Reader):
+        self.seen: list[ZarrLocation] = []
+        if isinstance(root, (Node, Reader)):
             self.seen = root.seen
         else:
-            self.seen = cast(List[ZarrLocation], root)
+            self.seen = cast(list[ZarrLocation], root)
         self.__visible = visibility
 
         # Likely to be updated by specs
         self.metadata: JSONDict = dict()
-        self.data: List[da.core.Array] = list()
-        self.specs: List[Spec] = []
-        self.pre_nodes: List[Node] = []
-        self.post_nodes: List[Node] = []
+        self.data: list[da.core.Array] = list()
+        self.specs: list[Spec] = []
+        self.pre_nodes: list[Node] = []
+        self.post_nodes: list[Node] = []
 
         # TODO: this should be some form of plugin infra over subclasses
         if Labels.matches(zarr):
@@ -62,14 +62,16 @@ class Node:
             self.specs.append(Well(self))
 
     @overload
-    def first(self, spectype: Type["Well"]) -> Optional["Well"]:
+    def first(self, spectype: type["Well"]) -> Optional["Well"]:
+        # Handled by the generic case
         ...
 
     @overload
-    def first(self, spectype: Type["Plate"]) -> Optional["Plate"]:
+    def first(self, spectype: type["Plate"]) -> Optional["Plate"]:
+        # Handled by the generic case
         ...
 
-    def first(self, spectype: Type["Spec"]) -> Optional["Spec"]:
+    def first(self, spectype: type["Spec"]) -> Optional["Spec"]:
         for spec in self.specs:
             if isinstance(spec, spectype):
                 return spec
@@ -100,7 +102,7 @@ class Node:
                 node.visible = visibility
         return old
 
-    def load(self, spec_type: Type["Spec"]) -> Optional["Spec"]:
+    def load(self, spec_type: type["Spec"]) -> Optional["Spec"]:
         for spec in self.specs:
             if isinstance(spec, spec_type):
                 return spec
@@ -225,7 +227,7 @@ class Label(Spec):
             LOGGER.warning("no parent found for %s: %s", self, image)
 
         # Metadata: TODO move to a class
-        colors: Dict[Union[int, bool], List[float]] = {}
+        colors: dict[Union[int, bool], list[float]] = {}
         color_list = image_label.get("colors", [])
         if color_list:
             for color in color_list:
@@ -235,7 +237,7 @@ class Label(Spec):
                     if rgba:
                         rgba = [x / 255 for x in rgba]
 
-                    if isinstance(label_value, bool) or isinstance(label_value, int):
+                    if isinstance(label_value, (bool, int)):
                         colors[label_value] = rgba
                     else:
                         raise Exception("not bool or int")
@@ -243,7 +245,7 @@ class Label(Spec):
                 except Exception:
                     LOGGER.exception("invalid color - %s", color)
 
-        properties: Dict[int, Dict[str, str]] = {}
+        properties: dict[int, dict[str, str]] = {}
         props_list = image_label.get("properties", [])
         if props_list:
             for props in props_list:
@@ -290,7 +292,7 @@ class Multiscales(Spec):
         # This will get overwritten by 'omero' metadata if present
         node.metadata["name"] = multiscales[0].get("name")
         paths = [d["path"] for d in datasets]
-        self.datasets: List[str] = paths
+        self.datasets: list[str] = paths
         transformations = [d.get("coordinateTransformations") for d in datasets]
         if any(trans is not None for trans in transformations):
             node.metadata["coordinateTransformations"] = transformations
@@ -350,9 +352,9 @@ class OMERO(Spec):
                 return  # EARLY EXIT
 
             colormaps = []
-            contrast_limits: Optional[List[Optional[Any]]] = [None for x in channels]
-            names: List[str] = [("channel_%d" % idx) for idx, ch in enumerate(channels)]
-            visibles: List[bool] = [True for x in channels]
+            contrast_limits: Optional[list[Optional[Any]]] = [None for x in channels]
+            names: list[str] = [("channel_%d" % idx) for idx, ch in enumerate(channels)]
+            visibles: list[bool] = [True for x in channels]
 
             for idx, ch in enumerate(channels):
                 # 'FF0000' -> [1, 0, 0]
@@ -383,7 +385,7 @@ class OMERO(Spec):
                     elif contrast_limits is not None:
                         contrast_limits[idx] = [start, end]
 
-            node.metadata["name"] = names
+            node.metadata["channel_names"] = names
             node.metadata["visible"] = visibles
             node.metadata["contrast_limits"] = contrast_limits
             node.metadata["colormap"] = colormaps
@@ -420,38 +422,34 @@ class Well(Spec):
         self.img_metadata = image_node.metadata
         self.img_pyramid_shapes = [d.shape for d in image_node.data]
 
-        def get_field(tile_name: str, level: int) -> np.ndarray:
+        def get_field(row: int, col: int, level: int) -> da.core.Array:
             """tile_name is 'row,col'"""
-            row, col = (int(n) for n in tile_name.split(","))
             field_index = (column_count * row) + col
-            path = f"{field_index}/{level}"
-            LOGGER.debug("LOADING tile... %s", path)
+            data = None
             try:
-                data = self.zarr.load(path)
+                # handle e.g. 2x2 grid with only 3 images/fields
+                if field_index < len(image_paths):
+                    image_path = image_paths[field_index]
+                    path = f"{image_path}/{level}"
+                    data = self.zarr.load(path)
             except ValueError:
                 LOGGER.error("Failed to load %s", path)
-                data = np.zeros(self.img_pyramid_shapes[level], dtype=self.numpy_type)
+            if data is None:
+                data = da.zeros(self.img_pyramid_shapes[level], dtype=self.numpy_type)
             return data
-
-        lazy_reader = delayed(get_field)
 
         def get_lazy_well(level: int, tile_shape: tuple) -> da.Array:
             lazy_rows = []
             for row in range(row_count):
-                lazy_row: List[da.Array] = []
+                lazy_row: list[da.Array] = []
                 for col in range(column_count):
-                    tile_name = f"{row},{col}"
                     LOGGER.debug(
                         "creating lazy_reader. row: %s col: %s level: %s",
                         row,
                         col,
                         level,
                     )
-                    lazy_tile = da.from_delayed(
-                        lazy_reader(tile_name, level),
-                        shape=tile_shape,
-                        dtype=self.numpy_type,
-                    )
+                    lazy_tile = get_field(row, col, level)
                     lazy_row.append(lazy_tile)
                 lazy_rows.append(da.concatenate(lazy_row, axis=x_index))
             return da.concatenate(lazy_rows, axis=y_index)
@@ -486,7 +484,6 @@ class Plate(Spec):
         LOGGER.info("plate_data: %s", self.plate_data)
         self.rows = self.plate_data.get("rows")
         self.columns = self.plate_data.get("columns")
-        self.first_field = "0"
         self.row_names = [row["name"] for row in self.rows]
         self.col_names = [col["name"] for col in self.columns]
 
@@ -502,6 +499,7 @@ class Plate(Spec):
         well_spec: Optional[Well] = well_node.first(Well)
         if well_spec is None:
             raise Exception("Could not find first well")
+        self.first_field_path = well_spec.well_data["images"][0]["path"]
         self.numpy_type = well_spec.numpy_type
 
         LOGGER.debug("img_pyramid_shapes: %s", well_spec.img_pyramid_shapes)
@@ -528,37 +526,38 @@ class Plate(Spec):
     def get_tile_path(self, level: int, row: int, col: int) -> str:
         return (
             f"{self.row_names[row]}/"
-            f"{self.col_names[col]}/{self.first_field}/{level}"
+            f"{self.col_names[col]}/{self.first_field_path}/{level}"
         )
 
     def get_stitched_grid(self, level: int, tile_shape: tuple) -> da.core.Array:
         LOGGER.debug("get_stitched_grid() level: %s, tile_shape: %s", level, tile_shape)
 
-        def get_tile(tile_name: str) -> np.ndarray:
+        def get_tile(row: int, col: int) -> da.core.Array:
             """tile_name is 'level,z,c,t,row,col'"""
-            row, col = (int(n) for n in tile_name.split(","))
+
+            # check whether the Well exists at this row/column
+            well_path = f"{self.row_names[row]}/{self.col_names[col]}"
+            if well_path not in self.well_paths:
+                LOGGER.debug("empty well: %s", well_path)
+                return np.zeros(tile_shape, dtype=self.numpy_type)
+
             path = self.get_tile_path(level, row, col)
-            LOGGER.debug("LOADING tile... %s with shape: %s", path, tile_shape)
+            LOGGER.debug("creating tile... %s with shape: %s", path, tile_shape)
 
             try:
+                # this is a dask array - data not loaded from source yet
                 data = self.zarr.load(path)
             except ValueError:
                 LOGGER.exception("Failed to load %s", path)
-                data = np.zeros(tile_shape, dtype=self.numpy_type)
+                data = da.zeros(tile_shape, dtype=self.numpy_type)
             return data
-
-        lazy_reader = delayed(get_tile)
 
         lazy_rows = []
         # For level 0, return whole image for each tile
         for row in range(self.row_count):
-            lazy_row: List[da.Array] = []
+            lazy_row: list[da.Array] = []
             for col in range(self.column_count):
-                tile_name = f"{row},{col}"
-                lazy_tile = da.from_delayed(
-                    lazy_reader(tile_name), shape=tile_shape, dtype=self.numpy_type
-                )
-                lazy_row.append(lazy_tile)
+                lazy_row.append(get_tile(row, col))
             lazy_rows.append(da.concatenate(lazy_row, axis=len(self.axes) - 1))
         return da.concatenate(lazy_rows, axis=len(self.axes) - 2)
 
@@ -568,7 +567,7 @@ class PlateLabels(Plate):
         """251.zarr/A/1/0/labels/0/3/"""
         path = (
             f"{self.row_names[row]}/{self.col_names[col]}/"
-            f"{self.first_field}/labels/0/{level}"
+            f"{self.first_field_path}/labels/0/{level}"
         )
         return path
 
@@ -587,10 +586,10 @@ class PlateLabels(Plate):
 
         # combine 'properties' from each image
         # from https://github.com/ome/ome-zarr-py/pull/61/
-        properties: Dict[int, Dict[str, Any]] = {}
+        properties: dict[int, dict[str, Any]] = {}
         for row in self.row_names:
             for col in self.col_names:
-                path = f"{row}/{col}/{self.first_field}/labels/0/.zattrs"
+                path = f"{row}/{col}/{self.first_field_path}/labels/0/.zattrs"
                 labels_json = self.zarr.get_json(path).get("image-label", {})
                 # NB: assume that 'label_val' is unique across all images
                 props_list = labels_json.get("properties", [])
@@ -619,7 +618,7 @@ class Reader:
     def __init__(self, zarr: ZarrLocation) -> None:
         assert zarr.exists()
         self.zarr = zarr
-        self.seen: List[ZarrLocation] = [zarr]
+        self.seen: list[ZarrLocation] = [zarr]
 
     def __call__(self) -> Iterator[Node]:
         node = Node(self.zarr, self)