+++ ome_zarr/scale.py
@@ -2,12 +2,13 @@
 
 See the :class:`~ome_zarr.scale.Scaler` class for details.
 """
+
 import inspect
 import logging
 import os
-from collections.abc import MutableMapping
+from collections.abc import Iterator, MutableMapping
 from dataclasses import dataclass
-from typing import Any, Callable, Iterator, List, Tuple, Union
+from typing import Any, Callable, Union
 
 import dask.array as da
 import numpy as np
@@ -25,7 +26,7 @@ from .io import parse_url
 
 LOGGER = logging.getLogger("ome_zarr.scale")
 
-ListOfArrayLike = Union[List[da.Array], List[np.ndarray]]
+ListOfArrayLike = Union[list[da.Array], list[np.ndarray]]
 ArrayLike = Union[da.Array, np.ndarray]
 
 
@@ -37,6 +38,21 @@ class Scaler:
     that do not begin with "_" and not either "methods" or "scale" are valid
     choices. These values can be returned by the
     :func:`~ome_zarr.scale.Scaler.methods` method.
+
+    Attributes:
+        copy_metadata:
+            If `True`, copy Zarr attributes from the input array to the new group.
+        downscale:
+            Downscaling factor.
+        in_place:
+            Does not do anything.
+        labeled:
+            If `True`, check that the values in the downsampled levels are a subset
+            of the values found in the input array.
+        max_layer:
+            The maximum number of downsampled layers to create.
+        method:
+            Downsampling method
 
     >>> import numpy as np
     >>> data = np.zeros((1, 1, 1, 64, 64))
@@ -76,9 +92,7 @@ class Scaler:
 
     def scale(self, input_array: str, output_directory: str) -> None:
         """Perform downsampling to disk."""
-        func = getattr(self, self.method, None)
-        if not func:
-            raise Exception
+        func = self.func
 
         store = self.__check_store(output_directory)
         base = zarr.open_array(input_array)
@@ -93,6 +107,14 @@ class Scaler:
             print(f"copying attribute keys: {list(base.attrs.keys())}")
             grp.attrs.update(base.attrs)
 
+    @property
+    def func(self) -> Callable[[np.ndarray], list[np.ndarray]]:
+        """Get downsample function."""
+        func = getattr(self, self.method, None)
+        if not func:
+            raise Exception
+        return func
+
     def __check_store(self, output_directory: str) -> MutableMapping:
         """Return a Zarr store if it doesn't already exist."""
         assert not os.path.exists(output_directory)
@@ -100,7 +122,7 @@ class Scaler:
         assert loc
         return loc.store
 
-    def __assert_values(self, pyramid: List[np.ndarray]) -> None:
+    def __assert_values(self, pyramid: list[np.ndarray]) -> None:
         """Check for a single unique set of values for all pyramid levels."""
         expected = set(np.unique(pyramid[0]))
         print(f"level 0 {pyramid[0].shape} = {len(expected)} labels")
@@ -115,7 +137,7 @@ class Scaler:
                 )
 
     def __create_group(
-        self, store: MutableMapping, base: np.ndarray, pyramid: List[np.ndarray]
+        self, store: MutableMapping, base: np.ndarray, pyramid: list[np.ndarray]
     ) -> zarr.hierarchy.Group:
         """Create group and datasets."""
         grp = zarr.group(store)
@@ -136,7 +158,7 @@ class Scaler:
         """
         if isinstance(image, da.Array):
 
-            def _resize(image: ArrayLike, out_shape: Tuple, **kwargs: Any) -> ArrayLike:
+            def _resize(image: ArrayLike, out_shape: tuple, **kwargs: Any) -> ArrayLike:
                 return dask_resize(image, out_shape, **kwargs)
 
         else:
@@ -154,7 +176,7 @@ class Scaler:
         )
         return image.astype(dtype)
 
-    def nearest(self, base: np.ndarray) -> List[np.ndarray]:
+    def nearest(self, base: np.ndarray) -> list[np.ndarray]:
         """
         Downsample using :func:`skimage.transform.resize`.
         """
@@ -165,7 +187,7 @@ class Scaler:
         if isinstance(plane, da.Array):
 
             def _resize(
-                image: ArrayLike, output_shape: Tuple, **kwargs: Any
+                image: ArrayLike, output_shape: tuple, **kwargs: Any
             ) -> ArrayLike:
                 return dask_resize(image, output_shape, **kwargs)
 
@@ -180,7 +202,7 @@ class Scaler:
             anti_aliasing=False,
         ).astype(plane.dtype)
 
-    def gaussian(self, base: np.ndarray) -> List[np.ndarray]:
+    def gaussian(self, base: np.ndarray) -> list[np.ndarray]:
         """Downsample using :func:`skimage.transform.pyramid_gaussian`."""
         return list(
             pyramid_gaussian(
@@ -191,7 +213,7 @@ class Scaler:
             )
         )
 
-    def laplacian(self, base: np.ndarray) -> List[np.ndarray]:
+    def laplacian(self, base: np.ndarray) -> list[np.ndarray]:
         """Downsample using :func:`skimage.transform.pyramid_laplacian`."""
         return list(
             pyramid_laplacian(
@@ -202,7 +224,7 @@ class Scaler:
             )
         )
 
-    def local_mean(self, base: np.ndarray) -> List[np.ndarray]:
+    def local_mean(self, base: np.ndarray) -> list[np.ndarray]:
         """Downsample using :func:`skimage.transform.downscale_local_mean`."""
         rv = [base]
         stack_dims = base.ndim - 2
@@ -211,7 +233,7 @@ class Scaler:
             rv.append(downscale_local_mean(rv[-1], factors=factors).astype(base.dtype))
         return rv
 
-    def zoom(self, base: np.ndarray) -> List[np.ndarray]:
+    def zoom(self, base: np.ndarray) -> list[np.ndarray]:
         """Downsample using :func:`scipy.ndimage.zoom`."""
         rv = [base]
         print(base.shape)